+++ System.Geometry.Text/Internal/Geometry.cs
@@ -63,7 +63,7 @@ namespace System.Geometry.Text
                         // if the RoundXYToGrid flag is set, round the offset components
                         if ((subglyph.Flags & CompositeGlyphFlags.RoundXYToGrid) != 0)
                         {
-                            offset = new Vector2((float)Math.Round(offset.X), (float)Math.Round(offset.Y));
+                            offset = new Vector2((double)Math.Round(offset.X), (double)Math.Round(offset.Y));
                         }
                     }
                     else
